[macOS] ScrollView's content are showing at End. (Style Cleanup)
